backend/requirements.txt
Flask
Flask-Cors
snowflake-connector-python
langchain
langchain-google-genai
python-dotenv
langchainhub
pandas
google-generativeai